Revert name change and remove central script
This file will now act as the central script
